backend/migrations/models/4_20230323195606_update.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `jobtype` RENAME TO `job_type`;"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `job_type` RENAME TO `jobtype`;"""
