backend/requirements.txt
aerich
tortoise-orm
sanic
python-dotenv